Fix ulong migration fallout
- Remove redundant casts and hoist checked conversions\n- Use ulong for semantically-unsigned nonce/gas fields and adjust tests\n- Make BlockDownloader.PrepareRequest swallow provider exceptions consistently

+++ src/Nethermind/Nethermind.Taiko/Rpc/TaikoEngineRpcModule.cs
@@ -212,7 +212,7 @@ public class TaikoEngineRpcModule(IAsyncHandler<byte[], ExecutionPayload?> getPa
                     }
 
                     // For Surge, filter out any transaction with very high gas limit
-                    if (surgeConfig.MaxGasLimitRatio > 0 && tx.GasLimit > tx.SpentGas * surgeConfig.MaxGasLimitRatio)
+                    if (surgeConfig.MaxGasLimitRatio > 0 && tx.GasLimit > tx.SpentGas * (ulong)surgeConfig.MaxGasLimitRatio)
                     {
                         worldState.Restore(snapshot);
                         while (i < txSource.Length && txSource[i].SenderAddress == tx.SenderAddress) i++;
